feat : add additional module packaging - feat: update `__init__.py` at root - feat: add `__init__.py` for `/sims` - feat: add `__init__.py` for `/plots` - feat: add `__init__.py` for `/core` - chore: update relative import - chore: move `cross_section.py` to `/sims` - chore: move `run_simulation.py` to `/sims` - chore: move `plotter.py` to `/plots` - chore: move `stopping.py` to `/core` - chore: move `helpers.py` to `/core` - chore: move `collision.py` to `/core` - chore: general linter pass - chore: rename `cross_section` to `monte_carlo`
