+++ src/strw_amuse/__init__.py
@@ -3,21 +3,15 @@ strw_amuse package for stellar dynamics simulations project using AMUSE.
 """
 
 from . import (
+    core,
+    plots,
+    sims,
     utils,
-    collision,
-    cross_section,
-    helpers,
-    plotter,
-    run_simulation,
-    stopping,
 )
 
 __all__ = [
+    'core',
+    'plots',
+    'sims',
     'utils',
-    'collision',
-    'cross_section',
-    'helpers',
-    'plotter',
-    'run_simulation',
-    'stopping',
 ]
