+++ src/strw_amuse/core/__init__.py
@@ -1,0 +1,11 @@
+"""
+strw_amuse.core package for stellar dynamics simulations project using AMUSE.
+"""
+
+from . import collision, helpers, stopping
+
+__all__ = [
+    'collision',
+    'helpers',
+    'stopping',
+]
