+++ src/strw_amuse/core/helpers.py
