+++ src/strw_amuse/plots/__init__.py
@@ -1,0 +1,9 @@
+"""
+strw_amuse.plots package for stellar dynamics simulations project using AMUSE.
+"""
+
+from . import plotter
+
+__all__ = [
+    'plotter',
+]
