+++ src/strw_amuse/sims/__init__.py
@@ -1,0 +1,10 @@
+"""
+strw_amuse.sims package for stellar dynamics simulations project using AMUSE.
+"""
+
+from . import monte_carlo, run_simulation
+
+__all__ = [
+    'monte_carlo',
+    'run_simulation',
+]
