+++ src/strw_amuse/utils/__init__.py
@@ -2,10 +2,10 @@
 strw_amuse.utils package for stellar dynamics simulations project using AMUSE.
 """
 
-from . import config, checker, logger
+from . import checker, config, logger
 
 __all__ = [
-    'config',
     'checker',
+    'config',
     'logger',
 ]
